EEPROMDevice : Applied Luke's code review comments
