+++ src/Common/EmuEEPROM.h
@@ -103,31 +103,4 @@ extern xboxkrnl::ULONG XboxFactoryGameRegion;
 }
 #endif
 
-#include "../CxbxKrnl/SMDevice.h"
-
-class EEPROMDevice : public SMDevice {
-public:
-	using SMDevice::SMDevice;
-
-	// SMDevice functions
-	void Init();
-	void Reset();
-
-	void QuickCommand(int read);
-	uint8_t ReceiveByte();
-	uint8_t ReadByte(uint8_t command);
-	uint16_t ReadWord(uint8_t command);
-	int ReadBlock(uint8_t command, uint8_t *data);
-
-	void SendByte(uint8_t data);
-	void WriteByte(uint8_t command, uint8_t value);
-	void WriteWord(uint8_t command, uint16_t value);
-	void WriteBlock(uint8_t command, uint8_t* data, int length);
-
-private:
-
-};
-
-extern EEPROMDevice* g_EEPROM;
-
 #endif // EMU_EEPROM_H
