+++ src/CxbxKrnl/CxbxKrnl.cpp
@@ -66,6 +66,7 @@ namespace xboxkrnl
 
 #include "PCIBus.h"
 #include "SMBus.h"
+#include "EEPROMDevice.h" // For EEPROMDevice
 
 PCIBus* g_PCIBus;
 SMBus* g_SMBus;
@@ -964,14 +965,13 @@ __declspec(noreturn) void CxbxKrnlInit
 	SetupXboxDeviceTypes();
 
 #define SMBUS_SMC_SLAVE_ADDRESS 0x20 // = Write; Read = 0x21
-#define SMBUS_EEPROM_ADDRESS 0xA8 // = Write; Read = 0xA9
 #define SMBUS_TV_ENCODER_ID_CONEXANT 0x8A // = Write; Read = 08B
 #define SMBUS_TV_ENCODER_ID_FOCUS 0xD4 // = Write; Read = 0D5
 
 	// Init Hardware
 	g_PCIBus = new PCIBus();
 	g_SMBus = new SMBus();
-	g_EEPROM = new EEPROMDevice();
+	g_EEPROM = new EEPROMDevice((uint8_t*)EEPROM);
 	g_SMBus->ConnectDevice(SMBUS_EEPROM_ADDRESS, g_EEPROM);
 	// TODO : Handle other SMBUS Addresses, like PIC_ADDRESS, XCALIBUR_ADDRESS
 	// Resources : http://pablot.com/misc/fancontroller.cpp
